backend/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.3
cupshelpers==1.0
defer==1.0.6
distro==1.3.0
distro-info===0.21ubuntu2
entrypoints==0.3
Flask==1.1.1
greenlet==0.4.15
httplib2==0.11.3
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
keyring==17.1.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.8.0
macaroonbakery==1.2.1
Mako==1.0.7
MarkupSafe==1.1.0
msgpack==0.5.6
neovim==0.3.0
netifaces==0.10.4
oauthlib==2.1.0
olefile==0.46
pexpect==4.6.0
Pillow==5.4.1
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.32.0
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt===1.8.5-ubuntu0.3
python-dateutil==2.7.3
python-debian==0.1.34
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
reportlab==3.5.18
requests==2.21.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
system-service==0.3
systemd-python==234
ubuntu-advantage-tools==19.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
virtualenv==16.7.9
wadllib==1.3.3
Werkzeug==0.16.0
xkit==0.0.0
